refactor: 분리되어 있던 테스트 컴포넌트를 스토리북 파일로 이동
